refactor(webview): fold the remote-control command into builtin-commands
The command is a one-line host action, which is what that file is for — its own
header says richer commands with their own UI stay in separate files. It also
sits in the Context section next to Resume conversation, so the file it lived
in held one class and an import nobody else used.

The error text loses its helper too: one call site and one ternary did not earn
a function, let alone one exported from core/commands into cv-app.
